dashboard/requirements.txt
fastapi==0.79.0
jinja2==3.1.2
deta==1.1.0
python-dotenv==0.20.0
